# docs/api-spec.yaml
openapi: 3.0.3
info:
  title: Median Code API
  description: |
    Production-ready FastAPI code generation and API entity management with Clerk JWT authentication.

    ## Authentication

    All endpoints require Bearer JWT authentication via Clerk. Include your Clerk JWT token in the Authorization header:

    ```
    Authorization: Bearer <your-clerk-jwt-token>
    ```

    ## API Overview

    This API provides two main categories of endpoints:

    ### 1. Entity Management (CRUD Operations)

    Manage your API design entities with full CRUD operations:

    - **Namespaces** (`/namespaces`) - Create and manage namespaces for organizing entities
    - **APIs** (`/apis`) - Create and manage API definitions that group endpoints and tags
    - **Types** (`/types`) - Retrieve primitive and abstract type definitions (read-only)
    - **Validators** (`/validators`) - Retrieve validator definitions (read-only)
    - **Fields** (`/fields`) - Create, read, update, and delete field definitions
    - **Objects** (`/objects`) - Create, read, update, and delete object definitions
    - **Tags** (`/tags`) - Create, read, update, and delete endpoint tags
    - **Endpoints** (`/endpoints`) - Create, read, update, and delete API endpoint definitions

    All list endpoints support filtering by `namespaceId` query parameter.

    ### 2. Code Generation

...

    #### Generated Application Structure

    The generated zip file contains:
...
    - `requirements.txt` - Python dependencies
...

...

...

...

...
servers:
...
security:
  - BearerAuth: []
...
paths:
...
  /fields:
    get:
      tags:
        - Fields
      summary: List all fields
      description: Retrieve all field definitions. Fields can be filtered by namespace.
      operationId: listFields
      parameters:
        - name: namespaceId
          in: query
          required: false
          schema:
            type: string
          description: Filter fields by namespace ID
      responses:
        '200':
          description: Successfully retrieved fields
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Field'
        '401':
          $ref: '#/components/responses/Unauthorized'
    post:
      tags:
        - Fields
      summary: Create a new field
      description: Create a new field definition.
      operationId: createField
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - namespaceId
                - name
                - type
              properties:
                namespaceId:
                  type: string
                  example: "namespace-user"
                name:
                  type: string
                  example: "email"
                type:
                  type: string
                  enum: [str, int, float, bool, datetime, uuid]
                  example: "str"
                description:
                  type: string
                  example: "User email address"
                defaultValue:
                  type: string
                  example: ""
                validators:
                  type: array
                  items:
                    $ref: '#/components/schemas/FieldValidator'
      responses:
        '201':
          description: Field created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Field'
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /fields/{id}:
    get:
      tags:
        - Fields
      summary: Get field by ID
      description: Retrieve a specific field by its ID.
      operationId: getField
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
          description: Field ID
      responses:
        '200':
          description: Successfully retrieved field
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Field'
        '404':
          description: Field not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          $ref: '#/components/responses/Unauthorized'
    put:
      tags:
        - Fields
      summary: Update field
      description: Update an existing field definition.
      operationId: updateField
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
          description: Field ID
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                  example: "updated_field_name"
                description:
                  type: string
                  example: "Updated description"
                defaultValue:
                  type: string
                  example: "new_default"
                validators:
                  type: array
                  items:
                    $ref: '#/components/schemas/FieldValidator'
      responses:
        '200':
          description: Field updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Field'
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Field not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          $ref: '#/components/responses/Unauthorized'
    delete:
      tags:
        - Fields
      summary: Delete field
      description: Delete a field. Cannot delete if the field is used in any objects.
      operationId: deleteField
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
          description: Field ID
      responses:
        '204':
          description: Field deleted successfully
        '400':
          description: Cannot delete field (used in objects)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                detail: "Cannot delete field: used in 3 objects"
        '404':
          description: Field not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /objects:
    get:
      tags:
        - Objects
      summary: List all objects
      description: Retrieve all object definitions. Objects can be filtered by namespace.
      operationId: listObjects
      parameters:
        - name: namespaceId
          in: query
          required: false
          schema:
            type: string
          description: Filter objects by namespace ID
      responses:
        '200':
          description: Successfully retrieved objects
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ObjectDefinition'
        '401':
          $ref: '#/components/responses/Unauthorized'
    post:
      tags:
        - Objects
      summary: Create a new object
      description: Create a new object definition.
      operationId: createObject
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - namespaceId
                - name
                - fields
              properties:
                namespaceId:
                  type: string
                  example: "namespace-user"
                name:
                  type: string
                  example: "User"
                description:
                  type: string
                  example: "User object definition"
                fields:
                  type: array
                  items:
                    $ref: '#/components/schemas/ObjectFieldReference'
      responses:
        '201':
          description: Object created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ObjectDefinition'
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /objects/{id}:
    get:
      tags:
        - Objects
      summary: Get object by ID
      description: Retrieve a specific object by its ID.
      operationId: getObject
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
          description: Object ID
      responses:
        '200':
          description: Successfully retrieved object
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ObjectDefinition'
        '404':
          description: Object not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          $ref: '#/components/responses/Unauthorized'
    put:
      tags:
        - Objects
      summary: Update object
      description: Update an existing object definition.
      operationId: updateObject
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
          description: Object ID
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                  example: "UpdatedObjectName"
                description:
                  type: string
                  example: "Updated description"
                fields:
                  type: array
                  items:
                    $ref: '#/components/schemas/ObjectFieldReference'
      responses:
        '200':
          description: Object updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ObjectDefinition'
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Object not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          $ref: '#/components/responses/Unauthorized'
    delete:
      tags:
        - Objects
      summary: Delete object
      description: Delete an object. Cannot delete if the object is used in any endpoints.
      operationId: deleteObject
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
          description: Object ID
      responses:
        '204':
          description: Object deleted successfully
        '400':
          description: Cannot delete object (used in endpoints)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                detail: "Cannot delete object: used in 2 endpoints"
        '404':
          description: Object not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Clerk JWT token from authentication flow
  schemas:
...
    ObjectDefinition:
      type: object
      required:
        - id
        - namespaceId
        - name
        - fields
        - usedInApis
      properties:
        id:
          type: string
          description: Client-generated unique identifier
          example: "object-1"
        namespaceId:
          type: string
          description: Namespace this object belongs to
          example: "namespace-global"
        name:
          type: string
          description: Object name (will become Pydantic model class name)
          example: "User"
        description:
          type: string
          description: Object description
          example: "User account object"
        fields:
          type: array
          items:
            $ref: '#/components/schemas/ObjectFieldReference'
          description: Fields that compose this object
        usedInApis:
          type: array
          items:
            type: string
          description: Array of endpoint IDs that use this object
          example: ["endpoint-1", "endpoint-2"]

    # ========================================================================
    # Object Field Reference
    # ========================================================================
    ObjectFieldReference:
      type: object
      required:
        - fieldId
        - required
      properties:
        fieldId:
          type: string
          description: Reference to Field.id
          example: "field-1"
        required:
          type: boolean
          description: Whether this field is required in the object
          example: true

    # ========================================================================
    # Field Definition
    # ========================================================================
    Field:
      type: object
      required:
        - id
        - namespaceId
        - name
        - type
        - validators
        - usedInApis
      properties:
        id:
          type: string
          description: Client-generated unique identifier
          example: "field-1"
        namespaceId:
          type: string
          description: Namespace this field belongs to
          example: "namespace-global"
        name:
          type: string
          description: Field name (will become Pydantic field name)
          example: "email"
        type:
          type: string
          enum: [str, int, float, bool, datetime, uuid]
          description: Primitive type name
          example: "str"
        description:
          type: string
          description: Field description
          example: "User email address"
        defaultValue:
          type: string
          description: Default value expression (Python code)
          example: ""
        validators:
          type: array
          items:
            $ref: '#/components/schemas/FieldValidator'
          description: Validators applied to this field
        usedInApis:
          type: array
          items:
            type: string
          description: Array of endpoint IDs where this field is used
          example: ["endpoint-1"]

    # ========================================================================
    # Field Validator
    # ========================================================================
    FieldValidator:
      type: object
      required:
        - name
      properties:
        name:
          type: string
          description: Validator name (references ValidatorBase.name)
          example: "max_length"
        params:
          type: object
          additionalProperties: true
          description: Validator parameters (optional)
          example:
            value: 255

    # ========================================================================
    # Validator Base
    # ========================================================================
...
    Error:
      type: object
      required:
        - detail
      properties:
        detail:
          type: string
          description: Error message describing what went wrong
          example: "Invalid API specification: endpoint path must start with /"
  responses:
    Unauthorized:
      description: Unauthorized - Invalid or missing Clerk JWT token
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            detail: "Not authenticated"